pyproject.toml
[project]
name = "ham-minon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.6.4",
    "pydantic>=2.12.3",
]
